+++ src/algorithms/local_search/operator.cpp
@@ -12,6 +12,10 @@ All rights reserved (see LICENSE).
 namespace vroom {
 namespace ls {
 
+std::string Operator::get_name() const {
+  return _name;
+}
+
 Gain Operator::gain() {
   if (!gain_computed) {
     this->compute_gain();