+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `location_index` key for unassigned tasks and each step, if provided (#625)
 - Shared target to makefile, for creating Position Independent Code (#617)
 - Exposing some internals for Python through compile flags (#640)
+- Stats on local search operators use for dev/debug purposes (#658)
 
 ### Changed